start adding delete

+++ src/EventStore.Core/Services/PersistentSubscription/PersistentSubscriptionService.cs
@@ -66,7 +66,7 @@ namespace EventStore.Core.Services.PersistentSubscription
         {
             Log.Debug("create subscription " + message.GroupName);
             //TODO revisit for permissions. maybe make admin only?
-            /*var streamAccess = _readIndex.CheckStreamAccess(SystemStreams.AllStream, StreamAccessType.Write, message.User);
+            var streamAccess = _readIndex.CheckStreamAccess(message.EventStreamId, StreamAccessType.Write, message.User);
 
             if (!streamAccess.Granted)
             {
@@ -74,7 +74,7 @@ namespace EventStore.Core.Services.PersistentSubscription
                                     ClientMessage.CreatePersistentSubscriptionCompleted.CreatePersistentSubscriptionResult.AccessDenied,
                                     "You do not have permissions to create streams"));
                 return;
-            }*/
+            }
 
             if (_subscriptionsById.ContainsKey(message.GroupName))
             {
@@ -105,6 +105,22 @@ namespace EventStore.Core.Services.PersistentSubscription
         public void Handle(ClientMessage.DeletePersistentSubscription message)
         {
             Log.Debug("delete subscription " + message.GroupName);
+            //TODO revisit for permissions. maybe make admin only?
+            var streamAccess = _readIndex.CheckStreamAccess(message.EventStreamId, StreamAccessType.Write, message.User);
+
+            if (!streamAccess.Granted)
+            {
+                message.Envelope.ReplyWith(new ClientMessage.CreatePersistentSubscriptionCompleted(message.CorrelationId,
+                                    ClientMessage.CreatePersistentSubscriptionCompleted.CreatePersistentSubscriptionResult.AccessDenied,
+                                    "You do not have permissions to create streams"));
+                return;
+            }
+            List<PersistentSubscription> subscribers;
+            if (!_subscriptionTopics.TryGetValue(message.EventStreamId, out subscribers))
+            {
+                
+            }
+
             message.Envelope.ReplyWith(new ClientMessage.DeletePersistentSubscriptionCompleted(message.CorrelationId,
     ClientMessage.DeletePersistentSubscriptionCompleted.DeletePersistentSubscriptionResult.Success, ""));
